requirements.txt
fastapi
uvicorn
pdfplumber
faiss-cpu
sentence-transformers
numpy
pandas
pydantic
pytest
